backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
ecdsa==0.19.0
bcrypt==4.0.1 
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.31.1
pytz==2024.2
